Working
Issue with hanging is because of conflicting requests between TiT and Sming tasks.
Although there are no asynchronous transactions in use, when a transaction
is active on the TiT task and is pre-empted by Sming it all goes wrong.
Simple solution is to use a sempahore.
This only fixes HSPI though, we need to serialise all TiT stuff into the
Sming task somehow.